Add content and styles for news page
I populated the page with some dummy content,
I will look for an easy way for the client to be
able to add/edit a news post without needing the
help from the developer (me). Worst case scenario,
it takes a miniscule amount of time to modify or
add a blog post.

**This commit:** (7-8-25)
Started:  6:00 PM.
Ended: 6:17 PM.
Length: 17 minutes.

**Total Project Time:**
2 hour 45 minutes.